+++ examples/ssr/src/pages/websocket-sse/_components/Chat.ts
