+++ examples/ssr/src/pages/websocket-sse/_components/Tab.ts
